Mongo storage now at 100% unit test coverage

+++ src/programy/storage/stores/file/store/learnf.py
@@ -34,20 +34,26 @@ class FileLearnfStore(FileStore, LearnfStore):
         return "%s%s%s.aiml" % (learnf_dir, os.sep, client_context.userid)
 
     @staticmethod
-    def create_learnf_file_if_missing(learnf_path):
+    def _write_xml_to_learnf_file(learnf_path):
+        dirpath = FileStore._get_dir_from_path(learnf_path)
+        FileStore._ensure_dir_exists(dirpath)
+        with open(learnf_path, "w+", encoding="utf-8") as file:
+            file.write('<?xml version="1.0" encoding="UTF-8"?>\n')
+            file.write('<aiml>\n')
+            file.write('</aiml>\n')
+            file.close()
 
+    @staticmethod
+    def create_learnf_file_if_missing(learnf_path):
         if os.path.isfile(learnf_path) is False:
             try:
                 YLogger.debug(None, "Creating new learnf file [%s]", learnf_path)
-
-                with open(learnf_path, "w+", encoding="utf-8") as file:
-                    file.write('<?xml version="1.0" encoding="UTF-8"?>\n')
-                    file.write('<aiml>\n')
-                    file.write('</aiml>\n')
-                    file.close()
+                FileLearnfStore._write_xml_to_learnf_file(learnf_path)
+                return True
 
             except Exception as excep:
                 YLogger.exception_nostack(None, "Error Writing learnf to %s", excep, learnf_path)
+        return False
 
     @staticmethod
     def write_node_to_learnf_file(client_context, node, learnf_path):
@@ -58,7 +64,8 @@ class FileLearnfStore(FileStore, LearnfStore):
             tree = ET.parse(learnf_path)
         except Exception:
             # Assume invalid aiml file, so remove it and start again with a fresh copy
-            os.remove(learnf_path)
+            if os.path.exists(learnf_path):
+                os.remove(learnf_path)
             FileLearnfStore.create_learnf_file_if_missing(learnf_path)
             tree = ET.parse(learnf_path)
 
@@ -106,18 +113,24 @@ class FileLearnfStore(FileStore, LearnfStore):
         child.append(xml_category)
         return child
 
+    def _write_learnf_to_file(self, client_context, category):
+        xml_node = self.create_category_xml_node(client_context, category)
+
+        learnf_path = self._get_storage_path()
+        self._ensure_dir_exists(learnf_path)
+
+        learnf_fullpath = self.create_learnf_path(client_context, learnf_path)
+
+        self.create_learnf_file_if_missing(learnf_fullpath)
+
+        self.write_node_to_learnf_file(client_context, xml_node, learnf_fullpath)
+
     def save_learnf(self, client_context, category):
         try:
-            xml_node = self.create_category_xml_node(client_context, category)
-
-            learnf_path = self._get_storage_path()
-            self._ensure_dir_exists(learnf_path)
-
-            learnf_fullpath = self.create_learnf_path(client_context, learnf_path)
-
-            self.create_learnf_file_if_missing(learnf_fullpath)
-
-            self.write_node_to_learnf_file(client_context, xml_node, learnf_fullpath)
+            self._write_learnf_to_file(client_context, category)
+            return True
 
         except Exception as exc:
             YLogger.exception_nostack(client_context, "Failed to save learnf", exc)
+
+        return False
